Enable mixed precision training for diffusion RL (#634)
1. Adopt MixedPrecisionPolicy of FSDP for diffusion RL training
2. Keep VAE always fp32 for stability

+++ cosmos_rl/policy/trainer/diffusers_trainer/nft_trainer.py
@@ -232,7 +232,7 @@ class NFTTrainer(DiffusersTrainer):
     ):
         """Log multimodal data to Weights & Biases (Wandb) for visualization."""
 
-        mm_datas_cpu = mm_datas.detach().cpu()
+        mm_datas_cpu = mm_datas.detach().to(torch.float16).cpu()
         num_to_log = min(15, len(mm_datas_cpu))
         rewards_cpu = rewards.detach().cpu()
 
@@ -252,9 +252,7 @@ class NFTTrainer(DiffusersTrainer):
         for out_idx, sample_idx in enumerate(sample_indices):
             out_path = os.path.join(mm_data_dir, f"{current_step}_{out_idx}.{ext}")
             if modality == "video":
-                video = (
-                    mm_datas_cpu[sample_idx].to(torch.float16).numpy()
-                )  # [T, C, H, W]
+                video = mm_datas_cpu[sample_idx].numpy()  # [T, C, H, W]
                 frames = (video.transpose(0, 2, 3, 1) * 255).astype(np.uint8)
                 imageio.mimsave(
                     out_path,
